Move dbdriver_mdl.py into drivers/mdl.py to separate driver from sensors
I'll be adding a sensor class for the mdl's power values and want to
have pymdl contain a separate path for drivers and sensors so databear
can scan for sensor classes once an environment variable for sensor
class path searching is added.
Also drivers|sensors mimics the structure in databear itself.