Don't quit on data errors.
If sensors are missing data in the databear database because they
aren't set up right or there's a bug in the databear getdata, getsensor,
etc. don't throw an exception in mdldisplay, just show "Missing data" etc.

The 2020-03-03 03:03 is probably not the right fix for not having any
timestamps, but fixes the crash for now.